requirements.txt
Django>=5.0
psycopg2-binary>=2.9
Pillow>=10.0
reportlab>=4.0
openpyxl>=3.1
factory-boy>=3.3
Faker>=20.0
django-widget-tweaks>=1.5
python-decouple>=3.8
django-crispy-forms>=2.1
crispy-bootstrap5>=2.0
channels>=4.0
channels-redis>=4.1
daphne>=4.0
redis>=5.0
celery>=5.3
django-celery-beat>=2.5

xhtml2pdf>=0.2.15
python-docx>=1.1.0
